se crean los controladores para mercancia

+++ src/main/java/com/example/hyperclBack/Repositorios/MercanciaRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MercanciaRepositorio extends JpaRepository<Mercancia, Long>{
+public interface MercanciaRepositorio extends JpaRepository<Mercancia, Integer>{
 }
